JS-1193 Order_state_transistion contains a result value (exit code)

+++ engine-kernelcpp/javaproxy/com__sos__scheduler__engine__kernel__event__CppEventFactory.h
@@ -52,7 +52,7 @@ struct CppEventFactory : ::zschimmer::javabridge::proxy_jobject< CppEventFactory
 
     static ::javaproxy::com::sos::scheduler::engine::data::event::AbstractEvent newLogEvent(jint p0, const ::zschimmer::javabridge::proxy_jobject< ::javaproxy::java::lang::String >& p1);
     static ::javaproxy::com::sos::scheduler::engine::data::event::AbstractEvent newOrderStateChangedEvent(const ::zschimmer::javabridge::proxy_jobject< ::javaproxy::java::lang::String >& p0, const ::zschimmer::javabridge::proxy_jobject< ::javaproxy::java::lang::String >& p1, const ::zschimmer::javabridge::proxy_jobject< ::javaproxy::java::lang::String >& p2, const ::zschimmer::javabridge::proxy_jobject< ::javaproxy::java::lang::String >& p3);
-    static ::javaproxy::com::sos::scheduler::engine::data::event::AbstractEvent newOrderStepEndedEvent(const ::zschimmer::javabridge::proxy_jobject< ::javaproxy::java::lang::String >& p0, const ::zschimmer::javabridge::proxy_jobject< ::javaproxy::java::lang::String >& p1, jint p2);
+    static ::javaproxy::com::sos::scheduler::engine::data::event::AbstractEvent newOrderStepEndedEvent(const ::zschimmer::javabridge::proxy_jobject< ::javaproxy::java::lang::String >& p0, const ::zschimmer::javabridge::proxy_jobject< ::javaproxy::java::lang::String >& p1, jlong p2);
     static ::javaproxy::com::sos::scheduler::engine::data::event::AbstractEvent newTaskEndedEvent(jint p0, const ::zschimmer::javabridge::proxy_jobject< ::javaproxy::java::lang::String >& p1, jint p2);
 
     ::zschimmer::javabridge::Class* java_object_class_() const;
